requirements.txt
imageio==2.5.0
matplotlib==3.1.1
numpy==1.16.2
opencv-python==4.1.0.25
Pillow==6.2.1
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
sklearn==0.0
tensorboard==1.14.0
tensorboardX==1.8
torch==1.2.0
torchvision==0.4.0
tqdm==4.35.0
pytorch-lightning==0.5.3.2
